move to dogs topic as migration, move knexfile
