+++ examples/todos/db/index.js
@@ -1,4 +1,4 @@
 const Knex = require('knex')
-const config = require('../knexfile')[process.env.NODE_ENV || 'development']
+const config = require('./knexfile')[process.env.NODE_ENV || 'development']
 
 module.exports = Knex(config)
